revamping GLiNER server
