+++ examples/deployment/gliner_server/pyproject.toml
@@ -1,0 +1,34 @@
+[project]
+name = "gliner-server"
+version = "0.1.0"
+description = "GLiNER FastAPI server for PII detection and entity extraction"
+readme = "README.md"
+requires-python = ">=3.10"
+license = "Apache-2.0"
+dependencies = [
+    "fastapi>=0.100.0",
+    "gliner>=0.2.0",
+    "pydantic>=2.0.0",
+    "torch>=2.0.0",
+    "uvicorn>=0.20.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0.0",
+    "httpx>=0.24.0",
+]
+
+[project.scripts]
+gliner-server = "gliner_server.server:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gliner_server"]
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+pythonpath = ["src"]
